fixed connection to mongo db (gracias zoe)

+++ api-series/src/main/java/com/dh/series/model/Chapter.java
@@ -19,7 +19,7 @@ public class Chapter implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    private Long chapterId;
+    private Integer chapterId;
 
     private String name;
